Add support for console.{debug,error,info,log,warn}.
