+++ cli/crates/cli/src/output/mod.rs
@@ -1,0 +1,1 @@
+pub mod report;
