+++ cli/crates/common/src/types.rs
@@ -8,6 +8,15 @@ pub enum LocalAddressType {
     Unspecified,
 }
 
+#[derive(Clone, Copy, Debug, serde::Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum ResolverMessageLevel {
+    Debug,
+    Error,
+    Info,
+    Warn,
+}
+
 impl LocalAddressType {
     #[must_use]
     pub const fn to_ip_v4(&self) -> Ipv4Addr {
